CT-217 Add ping.json End point
ping.json is needed for pingdom to check if the service is up
I had to update the config gem so that the config.rb initializer
could match the one used in Ask Tool. The rest is pretty much a mirror
of how we implemented this in the Ask tool.