CT-237 Refactor site prism test
We discussed this a while ago where adding additional css classes
to the markup was a bit excessive for just testing purposes. Misha
and I took a look at it and agreed that it would be best to reuse
aria-label and use css pseudo selectors to find the elements on the
page.

1. No additional css classes
2. Reusing aria-labels markup which is currently used for responsive
   table layouts and accessibility.