CT-323 Add Open/Close case tabs
This does not show/hide existing content on the page like traditional
tabs do nor does it use ajax to inject the additional content.

This has been inspired by Companies House usage of tabs and
an example is here https://beta.companieshouse.gov.uk/company/04278064

Later we may decide to make it smarter but then we need to consider
accessibility in more detail.