CT-352 Add Feedback
This creates the Feedback model which we will need to store
the feedback users submit. The model itself consists of just
a single jsonb column as we do not know what other fields
will be required and this implementation of Feedback is only
temporary.

A feedback form is added to each page each the login page.
It allows a user to type a message and to then submit their
feedback without leaving the page (Uses AJAX).

This commit does not include success/error messages.