CT-330 Add "Who it's With"
At first it was decided that we should be showing business units
but as we do not have business units implemented we have decided
to use the drafters email address as a clue to who the case is
currently owned by. If it is not assigned to a KILO then it
displays 'DACU' as the Business Management Team (BMT) dealing
with the case.

Later this implementation will be swapped out for Business Units.