Send assignment notification emails to drafter
When a case is assigned to a drafter, an email summary of the case
with a link to that case is sent to the drafter.

This PR also adds full name to the User model, and the user:dev_entries
and user:demo_entries have been updated accordingly.

To migrate the existing demo entries in the database, run
rake data:migrate:all