CT-416 store S3 key with attachments, not url
The bucket key seems to be a better way to refer to the attachment, as with the
URL there's a whole lot of converting back and forth between key and URL. This
way, internally we're driven just on the bucket and key and expose the public
URL only when needed. This allows us to clean up and fix some of the existing
code.