CT-474 don't let assigners see the uploaded files list
CT-474 specs out that assigners only see the uploaded files list after the case
has been marked as responded-to, but really what this means is hiding the
uploaded files list for the assigners: they'll see the uploaded files in the
Response Details section which will be added as part of CT-484 et al.

This commit includes some minor refactoring of specs like creating shared
examples and bubbling up fixtures that are used in multiple rspec contexts.