CT-472 Change case details link on case list page
Currently all users can click on a case number on the
case list page to see the case details page.

This Commit changes the behaviours of the case number link.

As an assigner, I should always be taken to the case details
page when I click on a case number form the case list.

As a drafter, If I click on a case which has not been
accepted or rejected, I should be taken to the Case
Accept/reject page. This is so I cannot see all the case details
before I decide to accept/reject and also to save me from
having to click on the "Accept/Reject" action button.

As a drafter, If I click on a case which been accepted then
it should take me to the case details page and I should not
see the "Accept/Reject" action button.