CT-472 Add pundit policy to case detail page
Assigners should always be able to view a cases detail
regardless of what state the case is in.

Only a draft who accepted the case can view a cases details
while its still in drafting state. If the case is assigned
to them but not been accepted, they should be redirected to
the Edit assignments page where they can decide to either
accept or reject the case.

If a drafter has not accepted the case or been assigned the case
they are trying to view, the system will return them to the case
list page and tell them that they cannot view that case.