CT-512 Status copy changes
This commit updates status that the user sees on the case list,
case detail and email subject.

The case list page:
"Status" is now labeled as "Waiting for" which will have a have
a step unless it is closed.

Case detail page:
If the case is still open then "Waiting for" and the state is shown.
Otherwise if the case is closed then "Waiting for" is replaced with
a "Status" of "Case closed"

The email subject needed to be changed as well so that it would
continue to read "waiting for acceptance"