CT-561 Create rake task to generate demo entries for Teams, Users and roles
Add rake task to create seed data for dev and prod:

  rake db:seed:prod
  rake db:seed:dev

This replaces the users:demo_entries rake task

The teams DACU, HR, LAA, HMCTS are set up, and three users added to each team.

Users are names with the initials DA, HR, HM, LA according to which team they
are in. Roles are manageer for DACU team members, and responder for all others.