CT-422 Add primary navigation
This adds a single link to the case list page. It will
later be used to help users navigate to other various
views that will be added.

The specs check each page for each user type to unsure
that its present and that the 'active' link is selected
when it is appropriate.