CT-649 remove "safe" triggers for case state machine
We used to have two triggers methods for each case state machine event, one
"unsafe" (e.g. assign_responder!) and one "safe" (e.g. assign_responder). This
was proving hard to reason about and required a lot of extra code, so we've
decided we're going to stick with just the "unsafe" version and expect event
triggers to raise in normal usage.