CT-323 Closed Case View
This PR provides a view of closed cases, and adds the closed cases link
to the Global Nav bar.

This PR contains a view spec for the Global Nav Partial.   I think this, along with GlobalNavManager spec means that we can do away with most of the Primary Navigation feature spec, retaining the bits that check that it's not visible at all on the login page.

In the Global Nav spec we test that the correct nav links are set up depending on role (or will do in the future), and in the view spec, we test that the right links are produced given a particular GlobalNavManager set up, and that the link is marked as active if the current_page?(url) returns true.

That should be enough, I think; we're testing the mechanism, and therefore don't have to test each and every possible combination.