CT-687 Introduce heading/sub-heading page headings
I reused the existing _header partial but extended it to
allow for an optional sub_header text to be added. I rolled
it out all the various pages and because of some of the refactoring
I was able to remove a partial called _case_page_heading as that
was no longer needed or used.

Next commit will include the BEM (Block/Element/Modifier css)