CT-687 Apply heading/sub headings to assignments
New Assignments page had a specific use case where it
had to determine whether the assignment was for a new case
or a assignment on an existing case that had previously been
rejected.

As part of this work it meant that we did not have to worry
about the copy of the button in either cases so I removed it
of the assignments helper. We still needed to know whether it
was new/existing case for the sub heading.