CT-645 refactor selecting radio buttons on SitePrism pages
When JS is enabled, the way we choose radio button inputs changes thanks to how
they are implemented in govuk_elements_form_builder. This cropped up in another
commit that's adding the "Take case on" button, but as that commit is getting
pretty big this commit is being done separately.