CT-623 - remove preview docs when attachmnet removed
Prior to this PR, when an attachment was removed, the original document was deleted from
S3 storage, but not the pdf preview version.

This PR also inludes a rake task to remove all the existing PDF versions that have been left behind
by attachment removals carried out before this change. It should be run once after deployment:

```
  rake data:migrate:delete_orphan_pdf
```