CT-700 Mailer test is not testing properly
As part of the previous commit, I discovered that the
mailer test was failing. On closer inspection I found that
the Assignment Mailer spec was using a case that was not in
the expected state. ATM the system does and should only use
this mailer when a case has been assigned to a KILO. In the
mailer spec it was using an unassigned case.