CT-670 Internal Deadliine
This PR ensures that an internal deadline is calculated for every case when it is created.

Because we don't want to display internal deadlines for non-flagged cases, the CaseDecorator
will return nil if the case is not flagged for approval.

After deployment, we need to run the following rake task to apply internal deadlines to all
existing cases:

```
    rake data:migrate:set_internal_deadline
```