CT-699 remove de-escalate JS and use controller actions to set view
Previously de-escalation was using JS to show/hide the Take case on /
de-escalate section based on the result of the result of the de-escalate
controller action. This approach started to have some problems when trying to
add these buttons to the case details view (CT-712) so it's been change to
behave like the Take case on button and to rely on the controller action to
return the HTML that needs to replace the section on the page.