CT-741 Move db/seeders/case_seeder to lib/cts/cases, prettier errors
lib/cts is more general than just being about case seeding, and cts should be
able to do more than just seed cases. Also, when moving the commands to a
sub-command "cases", Thor oddly seems to use the class name for generating the
help so with CaseSeeder as the sub-command, the help text actually listed the
sub-command as `cts case_seeder` instead of `cts cases`.