CT-711 Add "Show more/Less more" under case request
ATM a case is viewed in full but this leads to a poor user
experience when a user has already viewed the case and is familiar
with it.

On the case details page when a request is more than 350 characters
the first 350 characters are shown along with the remaining text
being hidden until the user clicks "Show more".

If JS is disabled then the full request is shown.