CT-748 Default the home page based on the users role
This fixes the bug that was discovered where when the user
logs in and they are taken to the cases index page a tab is not
set by default. This also means that the case list is incorrect
as it displays all cases regardless of whether they are in time
or late.