iTimes were always being displayed in UTC which coincides with local time in London during the winter, but is one hour out when BST is in forc
This is fixed by setting the rails configuration time zone to London.

Times in the database are still recorded as UTC, but rails will automatically
convert them to the local timezone when instantiating the object, meaning that
the correct time is displayed winter or summer.

Have also stubbed out the call to the govuk site to get holidays for all environments other
than production so that tests and server can be run offline.