CT-803 Fix bug where not all attachments converted to PDF
This was caused by the PDF job firing off before the transaction to save the
CaseAttachment records was complete - most commonly seen when a lot of files
are uploaded in one go.  This fix removes the error handling from the Job, and
allows sidekiq to reschedule the job in it's own time.  This is recommended behaviour
for ActiveJob using sidekiq.