CT-823 Open Foi Cases report (stats)
There were originally two who_its_with methods: on Case, and on CaseDecorator.  These were
both out of date, and did different things.  This PR refactors them as follows:

Case#who_its_with replaced by Case#current_team_and_user which uses a service
to determine the teams (and if appropriate the user) that the case is currently
with.  The service raises an exception if a case is given to it is in an unrecognised
state - this prevents it becoming out of date as we add new states.

CaseDecorator#who_its_with refactored to use Case#current_team_and_user and decide whether
the team name and / or the user name is to be displayed.

Remove useless private declaration

Refactor Stats collector to use rows and columns and produce CSV

WIP

Use only FOI cases

Add report generator

Add report R002

Fix report to only take cases with no responded event

update description