CT-815 Add migration to add DACU Disclosure assignments.
As a result of CT-815, or the lack of this change actually, cases have been
escalated to require Press Office clearance without having DACU Disclosure also
assigned to them, which is required. This migration adds DACU Disclosure
assignments to any case that has been escalated to Press Office and that doesn't
have an existing DACU Disclosure assignment. This migration is idemnipotent.