CT-838 Use Closed Status to determine whether case has been responded for Stats report
Formerly, the presence or absence of a respond event transition determined whether or not
a case was considered closed for stats purposes.  Now we use the status of 'closed'.

Also, report was renamed from current month case overview to Business Unit Performance report
as this is the terminology they currently use