CT-811 Private Office take on case
This commit makes a start on taking on a case for private office.

Login as Private Office user and you should be able to view
new cases page and actions to take on a case.

This still needs to get the "Undo" working and needs to assign
to a specific user in Press Office. It already flags it for DACU
DS clearance (i.e making it a trigger case)