CT-812 event transitions can have 'policy' argument
Building on the per-transition guards, we can now specify a policy to check
before a transition is run. When triggering the event, the generated guard will
instantiate a policy object using the object and user id and call the given
policy method on it to check if the transition is permissable.

Also in this commit:

- Complete the tests for Events and EventTransitions
- Move permitted_events from CaseStateMachine to Events