CT-812 allow press officers to upload response for redrafting
Uses the new per-transition guard/policy mechanism, and also names the policy
methods after the controller action / state machine event, which is the
recommended way to use Pundit. The reasoning for this change is that it's been
really hard to match up the controller action or state machine events with
policies. e.g. there are certain actions and events, such as uploading
attachments to files, which require many policies, it just doesn't seem
practical to have them re-use policies. The new thinking is we're better off
making it easy to find the relavant policy method and making the check methods
the units of re-use.