CT-840 pre-populate date responded for managers
On the case closure page, DACU BMT (i.e managers) would prefer we
pre-populate the date responded fields with the date the KILO (Responder)
clicked on the "Mark as sent". This gives them the option to modify it
if its different from the date the KILO actually sent the response out.