CT-859 don't create extra business units in reports tests
Don't know why but this was suddenly causing error, somehow the line:

  Team.where.not(id: [@team_1.id, @team_2.id]).map(&:destroy)

was deleting all the teams, even though the where appeared to match only the
right ones. A cleaner way to prevent all this noise of too many teams is just to
use the responder when creating cases so the fixtures don't create an extra
responding_team.