CT-1120 Add transitions to add_message_to_case
Resolves an issue raise by Sentry where if a cases is in a specific
state the users get an error message. The reason is that when
we added messages feature we did not have press/private office
transitions in the workflow.

The initial error was raise for cases pending press clearance but
having looked at the code it would appear that private office would
have the same issue