CT-1124 Bug - users cannot download Business Unit Performance Report
This bug was caused by a user changing the role of a team from responding
to approving.  The Report only lists responding teams, so that when it encounters
a case which is allocated to that team, it cannot find a row to put it against.

The immediate fix is to show all teams, regardless of role, but we need
to put some work around the editing for roles.