CT-1060 - Add permance percentage columsn to R003
This PR adds three performance percentage reports to R003 -
Business Unit Performance Report as well as the name of the
team lead in each team.

This work is done by adding the ability to accept callback methods to the
Stats generator which allows the hierarchical rollup of statistics to be
accomplished, as well as addition of team lead roles.