CT-1034 Add mailer to notify press/private user
This adds an Assignment method that will eventually be used by the
approval service. An assignment is passed to the method it then extracts
the case details and the user who took on the case, both used for the
govuk notify templates placeholders and the recipents email address.

I had to create an GOVUK notify template first and outside the code
base, then copy the template id which you will in the settings.yml

I modified the approval service to hook into the mailer so that it
would only be called when a case is pending press/private clearnance
and doesn't execute when the clearance is bypassed.