CT-602 remove smoketests need for db interation
Initial we were going to need to check and add smoke test
users and cases but after numerous issues with setting up jenkins to try
read DB username/Password and other config issues we decided that maybe
we should just add smoke test user to dev seed users and manually
create it in production. if something does happen to the user account
the worst that could happen is that the smoketests would fail.