CT-1106 comments out team deactivation links
As we do not have a method to deactivate business units the directorates and business groups cannot be deactivated we have therefore removed the link to deactivate business groups and directorates to avoid user confusion.

Once we have the functionality to deactivate business units we can uncomment this code and use it again.