CT-1301 Allow all users to view any case details
This ticket opens up the case details to allow anyone to view the
details of the case but still restrictive enough that only people
authorised to take action on the case can still do so.

This is to help enable the user story that relates to users being
able to link cases to each other.