CT-1387 fixes failing smoketest
The tests were failing because the structure of the markup had changed.

This fix makes the tests just check if the case details page heading
contains the case number and the summary of the case that was
clicked instead of worrying about extra whitespace or additional content
being added to the headings like what type of case it is.