CT-1391 remove category association in case
Case is now an STI model, so we rely on the 'type' field now and the category
just supplies additional information like deadlines, etc. We link the case to a
category using the 'type_abbreviation' (FOI, SAR, etc) defined statically in the
class (could also be derived from 'type') so that we can retrieve information
about the category that way.