CT-1391 rename Case::SAR::NonOffender to Case::SAR
After a discussion with Mara it became evident that SAR::NonOffender is a
misnomer as these cases can involve an offender, it's just that those requests
about an offender involving prison or probation are dealt with by another team
and need another workflow. The second type we were going to call SAR::Offender.

Until we get to doing those case types, it's probably best to just have one
type, SAR, and we can always have a sub-type that is called something more
appropriate once we figure that out.