CT-1476 BUGFIX Prevent add_message_to_case being called with wrong team
This is a temporary fix to solve the problem of calling an event on a state machine
and passing in an acting team which doesn't have the right to execute that event.

In this case, it is the messages controller, which determines the acting team with:

team = current_user.teams_for_case(@case).first

In the old code, this would have returned one of the teams that was common across
both the case and the user.  If it was the Disclosure team
first, rather than the Disclosure BMT team, then an error occurs because the state
machine rightly says that the action isn't allowed for Disclosure team.

This is a temporary fix to ensure that the messages controller passes in the first
team after they have been ordered in priority 'manager', 'approver', 'responder', which
will work in this case.

But we need to find a more generalised and permanent solution - perhaps asking the
state machine which of the teams has this right.