CT-1391 make third_party and reply_method validations better
Further clean-up and improvements to validations and how we present them. Here
we add the gem 'humanize_boolean' to tidy up how we present 'third_party' values
and allow 'reply_method' to be presented as 'send_by_email' and 'send_by_post'
instead of 'email' and 'post'. By doing this, we remove the translations
third_party and reply_method in en.helpers.label.case_sar, which were blocking
presenting the validation errors correctly: when the values for these
translations are a hash, as was the case, the name of the attribute in the
validation error text is stripped out by the UK Gov FormBuilder (see PR 67
https://github.com/ministryofjustice/govuk_elements_form_builder/pull/67).