CT-1391 move JSON attributes into Case::FOI::Standard and Case::SAR
The correct way to use these JSON attributes is to not have them show up on
every case, by putting them in Case::Base, but to have them specific to the
different sub-classes of cases. This does mean that common attributes have to be
redefined in multiple places, but this is probably ok as attributes common to
all cases should probably be defined as columns anyway.