CT-1657 Non-trigger FOI in awaiting responder use Configuratble State Machine
This PR changes the awaiting_responder state for non-trigger FOI cases to use
the configurable state machine.  It also introduces the concept of different
workflows for trigger and non trigger cases, which is recorded on the case and
on each transition.

There is a temporary test which tests the state, the state machine and transition
after every action in the end-to-end user journeys which we will use while
converting the other states to use the configuratble state machine.