CT-1571 add support for uploading test failure screenshots to S3
Every now and then we have sporadic failures in Travis that much upset and
gnashing of teeth. This enables a mechanism built-into capybara-screenshots to
upload sceenshots of errors directly to S3 from the Travis host for our viewing
pleasure.

To enable this feature, env vars have to be set in Travis for the access key and
secret access key of a user that has access to a bucket we've set aside for this
purpose. File uploaded to this but will be expired after 8 weeks.