CT-1243 ability to add deactivated users to a team
ATM, when you try to add a deactivated user to a team. It reports
that an existing user has been added to the team but it does not list
them in the teams users list. The reason for this is because the user
is still soft deleted.

This commit removes the soft delete from the users account
if the user is deactivated