CT-1609 Update Search Index
Every time a Case::Base record gets updated, if any of the amended fields those
used in the search index, the dirty flag is set and a SearchIndexUpdater
job is scheduled to run after a delay of 10 seconds.

This also happens if an responding assignment is created or updated.

The search index updater job will collect all records in the database with the
dirty flag set, and call SearchIndex#update_document on them.

The reason for the 10 second delay is to prevent the job running multiple times
on the same record if the same record is updated several times during a transaction
(assignments to responding team will also cause the dirty flag to be set).