CT-1614 Do not show remove clearance link for press and private office users
This bug was caused by the conflation of ‘unflag for clearance’ and ‘unaccept_approver_assignment’ (press and private officers use the former when they should be using the latter).  There is a ticket to refactor this, but in the meantime, the view calls the policy to check whether or not to show the link.