CT-1599 Record search queries and clicks
This PR iplements the recording of search queries and clicks on results
to enable an alysis of the effectiveness of full text search.

It is implemented in the app rather than Google Analytics because of the
sensitive nature of some of the search terms.

Every time a search is made, a hash is calculated from the user id, date
and search query and a record written to the search queries table along
with the number of results.  The links to the case number displyed on the
search results page have the hash of the search query and position within
the result set embedded within them.

When a case is displayed, if the URL had a search query hash embedded in it,
it records the click on the table.

Currently, clicking on a search result and then reloading the page will cause
2 clicks to be recorded.  This will be addressed in a later iteration.