CT-1672 Record filters in search_queries table for search/filtering metrics
Filtering is usually carried out on a result set obtained as a result of a
search.  We don't want the search to be recorded as zero clicks, but rather
indicate that a filter was carried out on the search results (and maybe a
subsequent filter performed on that result set), and then record the number
of clicks against that.

This is acheived by writing a search query record which has a parent_id
relating it to the preceeding search or filter.

All of this is controlled by the CaseSearchService, which works out whether
or not this is child filter, and writes the appropriate record, performing t
the original search and then passing the arel thus obtained to
the CaseFilterService for filtering, before finally passing the filtered
result set back to the controller for display.

This PR implements a dummy filter action on the cases controller, which will
be replaced with the real filter when filtering is implemented.