CT-1682 Allow view-only scope for responders
Up to now, the policy scope returned for users who are only responders
has comprised cases that were assigned to their team.  However, now we
need a new scope to encompass all Cases that can be used when searching.

This is acheived by adding the method #for_view_only on Case::BasePolicy::Scope
and instantiating that manually in the case_search_service rather than relyng on
the Pundit helper method.

This PR also renames ```case_policy_spec.rb``` to ```base_policy_scope_spec.rb```
in order to more accurately reflect what it is testing, and create the cases in a before(:all)
scope in order to reduce run time of the spec.