CT-1207 make CaseSearchService find previous search queries properly
This functionality is crucial replaying already-selected filters and for
attributing stats to the correct search query for metrics. We now find an
existing search query by matching on search parameters (search test, filters,
etc), user and date, and present that query back to the user instead of creating
a new one.