CT-1671 fix "Take on" action on the case details page
Users reported that that "Take on" button was not working on the case details
page. When I investigated this issue I discovered that it was taking the
case on but it was not displaying any changes to the page until the Users
refreshed the page.

The functionality to take/de-escalate/undo was copied between the "New cases"
case list page and the case details page. It relied purely on an AJAX request
which would work on the case list but was not designed to work on the case
details page. I did not just update the javascript as there was a lot of
other markup and functionality that the user would need to see and use
compared to the case list which is mainly used to take on cases in "Bulk".

As a solution to the problem, I added a `respond_to` to the action to either
render the actions javascript if it was an ajax request or redirect it to
the case details page if it was a normal HTTP request. If it is from the
redirect then we add a flash message confirming the change on long with
a link to the undo action.

While I was investigating, I discovered that Press/Private Office Users
cannot take on cases from the case details page. We should investigate
this as a separate ticket.