CT-1621 makes request to create_from_s3
now when you upload a file, the file is uploaded to S3. Once it is finished,
dropzone calls the CaseAttachment#create_from_s3 action and passes in a
path to the file in the S3 bucket. The action queues the virus scan job
and then returns JSON with a location to CaseAttachment#show. Once Dropzone
recieves success and JSON it then does an ajax request to the show action.
and does this once 1 sec until the file state changes.

All this works except for the virus scanner actually passing a file.