CT-1459 Tidy up show? policies for FOI standard and SAR cases
This change ensures that responders cannot view SAR cases (this
will be changed when we introduce trigger SARS to allow disclosure
specialist to view trigger SAR cases,  but not press/ private.

It also replaces the #show? method on the Case::BasePolicy to not allow
anyone to virew.  This is handled by the policies for Case::FOI::Standard
and Case::SAR