CT-1405 Allow users to see closed SAR cases
In fact, all users that need to see closed SAR cases (BMT and responders)
can see them, but cannot distinguish between SAR cases and FOI cases.  This PR
adds the Type column on the closed case list to show the type (FOI or SAR) as well
as the trigger badge.