CT-1741 press and private officers are now in CorrespondenceType
We didn't have CorrespondenceType with jsonb attributes when we first added
press and private default users, so we stuck it in the settings.yml with ENV
overrides in the deployed environment That's too much maintenance so this gives
us a simpler way to do manage this in the db. We haven't exposed this in the GUI
yet, but that would be the next logical step.