CT-1766 Make Business unit columns depend on parameter in Business Unit report
The business unit columns were added to this report in order to surface cases where
the business units were responding in time, but the case was late due to other reasons.

This proved to be controversial, and so we have removed it from the report that is
generated from the stats page by making the appearance/non-appearance depend on a parameter.