CT-1542 Implement Business Unit performance report for SARs
This is implemented by moving the logic from the existing
R003BusinessUnitPeformanceReport into a new class BaseBusinessUnitPerformanceRepoert
and deriving the Business unit performce reports for both FOIs and SARs from that.  The
derived classes just differ in returning different names nad desciptions of the report, and
most importantly the case_scope, which determines which cases will be reported on

The specs for the new R103SarBusinessUnitPerformanceReport only exercise the new methods; the
code in the base class for analysing, totalling, calculating percentages in thoroughly testing in
the spec for the FOI report.