CT-1722 Refactor SAR close to respond_and_close
SAR cases do not have uploaded responses, and so the first iteration was
that the responder simply closed the case when ready.  However, this meant
that the way of testing the response date of a case differed between FOI and
SAR cases, in that there was no respond event for SAR cases, and the close
event had to be used a a proxy.

This PR changes the close event for SARs to a respond_and_close which in fact
writes two events tothe transitions table: respond immediately followed by a
close