CT-1793 Add SAR Monthly report
This is acheived by moving most of the code from the R005MonthlyPerformanceReport
into a BaseMonthlyPerformanceReport class, and deriving the original R005 report
for FOIs and the new R105 report for SARs from it.

Test for the new R105SarMonthlyPerformanceReport only exercise the methods in the
derived class, .title, .desdcription and #case_scope; all the methods in the base
class are fully exercised inthe spec for R005.