chrome does not find non-visible text
A new matcher called have_copy was created to find content on the page
regardless whether its hidden or not. It also solves the problem where
we use non-visible text to add context for screenreaders.

The new matcher turns on/off the flag that instructs Capybara to ignore
hidden elements/text.

https://makandracards.com/makandra/7617-change-how-capybara-sees-or-ignores-hidden-elements