CT-1776 Refactor the views and functionality
After speaking to Mia, We decided to refactor the designs and only allow
users to link to original case. The ajax call now returns cases linked
to the original case. Users have the option to remove any other related
cases that do not involve the current ICO.

We need the new creation ICO vieE and tests to be written.