CT-1824 Bugfix for responders not able to upload after approval
If a responder has uploaded a document to a flagged case, and that case
has been approved by disclosure, but amends requested by Press or Private,
the responder was not able to upload documents again once the case was
back in drafting due to a condition that prohibited it if there was an approved
approver assignemtn on the case.  This fix removes that prohibition, allowing
responders to upload if the case is in drafting regardless of whether or not the
case has previously been approved.